Implemented the new source in the mixer, and improved the mixer code quite a bit

+++ components/audio_renderer/audio_renderer.h
@@ -22,7 +22,7 @@ typedef struct {
 void renderer_init(renderer_config_t *config);
 void renderer_destroy();
 
-int render_samples(int16_t *buf, size_t buf_len);
+int render_samples(uint8_t *buf, size_t buf_len);
 void renderer_set_sample_rate(uint32_t sample_rate);
 void renderer_clear_dma();
